==> 13-germany-mods-clean.Rmd <==
---
title: "19F in Germany - spatiotemporal models in INLA"
author: "Paloma Cárcamo"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
if (!require("pacman")) install.packages("pacman")
pacman::p_unload(pacman::p_loaded(), character.only = TRUE)
# install.packages("INLA",repos=c(getOption("repos"),INLA="https://inla.r-inla-download.org/R/stable"), dep=TRUE)
pacman::p_load(tidyverse, sf, spdep, INLA, cowplot, biscale)
```

## Load data

Shapefiles from GADM https://gadm.org/download_country.html

Vaccination coverage estimates from Robert Koch Institut https://robert-koch-institut.github.io/Inanspruchnahme_von_Routineimpfungen_in_Deutschland-Ergebnisse_aus_der_KV-Impfsurveillance/

```{r}
# IPD case data
ipd_raw <- read_csv("data/DE_IPD_deidentified.csv")

# shapefile
shp <- read_sf("data/plz-2stellig.shp") |> 
  group_by(plz) |> 
  summarise(geometry = st_union(geometry))

# vaccine coverage
vax_cov <- readxl::read_xlsx("data/kvis_vax.xlsx")
```

## Link case dataset to vaccine coverage dataset

(this takes a while to run)

```{r}
# ger_shp <- read_sf("data/gadm41_DEU_2.shp")
# 
# ger_shp_j <- ger_shp |>
#   mutate(ags_area = st_area(geometry))
# 
# st_intersection_faster <- function(x, y) {
#   y_indices <- st_intersects(x, y) |>
#     unlist() |>
#     unique() |>
#     sort()
# 
#   y_subset <- y[y_indices, ]
#   st_intersection(x, y_subset)
# }
# 
# intersection_pre <- st_intersection_faster(ger_shp, shp)
# 
# intersection <- intersection_pre |>
#   mutate(intersection_area = st_area(geometry),
#          district_area = ger_shp_j$ags_area[match(intersection_pre$CC_2, ger_shp$CC_2)],
#          overlap_percentage = (intersection_area / district_area) * 100)
# 
# plz_ags_key <- intersection |>
#   st_drop_geometry() |>
#   dplyr::select(CC_2, plz, overlap_percentage) |>
#   arrange(CC_2, plz)
# 
# # write_rds(plz_ags_key, "data/interim/plz_ags_key.rds")

plz_ags_key <- read_rds("data/interim/plz_ags_key.rds")
```

## Join all datasets

```{r}
# aggregate PCV vaccine coverage data by 2-digit zipcode and by 4-year period
vax_cov_agg <- vax_cov |> 
  filter(vaccine == "Pneumokokken" & age_group == "24 Monate") |> 
  left_join(plz_ags_key, by = c("county_ags" = "CC_2")) |> 
  filter(!is.na(plz)) |> 
  mutate(new_pop = pop_weight*overlap_percentage,
         vaxed = new_pop*(vacc_rate/100),
         year = year_birth+2,
         nyeargroup = cut(year, breaks = c(2008, 2012, 2017, 2022), include.lowest = TRUE, labels = FALSE)) |> 
  group_by(plz, nyeargroup) |> 
  summarise(vax_pop = as.numeric(sum(vaxed, na.rm = TRUE)),
            all_pop = as.numeric(sum(new_pop, na.rm = TRUE))) |> 
  ungroup() |> 
  mutate(vacc_rate = vax_pop/all_pop,
         vacc_cat = inla.group(vacc_rate, n = 10, method = "quantile")) |> 
  dplyr::select(-vax_pop, -all_pop)

# DPT 1st dose vaccine coverage (proxy for PCV booster)
dpt <- vax_cov |> 
  filter(vaccine == "Diphtherie, Tetanus, Pertussis") |> 
  left_join(plz_ags_key, by = c("county_ags" = "CC_2")) |> 
  filter(!is.na(plz)) |> 
  pivot_wider(names_from = vaccination_status, values_from = vacc_rate) |> 
  mutate(new_pop = pop_weight*overlap_percentage,
         vaxed_1d = new_pop*(`1. Dosis`/100),
         vaxed_3d = new_pop*(`3. Dosis`/100),
         year = year_birth+1,
         nyeargroup = cut(year, breaks = c(2008, 2012, 2017, 2022), include.lowest = TRUE, labels = FALSE)) |> 
  group_by(plz, nyeargroup) |> 
  summarise(dpt_vax1d_pop = as.numeric(sum(vaxed_1d, na.rm = TRUE)),
            dpt_vax3d_pop = as.numeric(sum(vaxed_3d, na.rm = TRUE)),
            dpt_all_pop = as.numeric(sum(new_pop, na.rm = TRUE))) |> 
  ungroup() |> 
  mutate(dpt_vacc_rate = dpt_vax1d_pop/dpt_all_pop,
         dpt_vacc_cat = inla.group(dpt_vacc_rate, n = 10, method = "quantile")) |> 
  select(plz, nyeargroup, dpt_vacc_rate, dpt_vacc_cat)
  
# join vaccine and IPD datasets
pcv7_st <- c("4","6B","9V","14","18C","19F","23F")
pcv7_no19f <- c("4","6B","9V","14","18C","23F")

data_full <- ipd_raw |> 
  mutate(date = as.Date(DateOfIsolation, format = "%m/%d/%Y"),
         plz_pre = str_pad(PLZpatient, side = "left", pad = "0", width = 3),
         plz = substr(plz_pre, 1, 2),
         year = year(date)) |> 
  filter(!is.na(PLZpatient) & date > as.Date("2009-01-01") & !is.na(date) & !(plz %in% c("05", "11"))) |> 
  arrange(year) |> 
  group_by(year) |> 
  mutate(nyear = cur_group_id()) |> 
  ungroup() |> 
  mutate(yeargroup = factor(cut(nyear, 3, labels = FALSE), levels = c(1:3), labels = c("2008-2012", "2013-2017", "2018-2022")),
         nyeargroup = cut(nyear, 3, labels = FALSE)) |> 
  group_by(plz, Serotype, yeargroup, nyeargroup) |> 
  summarise(ipd_st = n()) |> 
  group_by(plz, yeargroup, nyeargroup) |> 
  mutate(ipd = sum(ipd_st),
         ipd_pcv7 = sum(ipd_st[Serotype %in% pcv7_no19f], na.rm = TRUE)) |> 
  ungroup() |> 
  filter(Serotype %in% pcv7_st | Serotype == "19A") |> 
  pivot_wider(names_from = Serotype, names_prefix = "ipd_", values_from = ipd_st) |> 
  mutate(across(starts_with("ipd_"), ~ if_else(is.na(.x), 0, .x)), 
         across(starts_with("ipd_"), ~ .x / ipd, .names = "ratio_{.col}"),
         across(starts_with("ratio_ipd_"), ~ log((.x + 0.001) / (1 - .x + 0.001)), .names = "logit_{.col}")) |> 
  left_join(vax_cov_agg, by = c("plz", "nyeargroup")) |> 
  group_by(yeargroup) |> 
  mutate(meancov = mean(vacc_rate, na.rm = TRUE)) |> 
  ungroup() |> 
  mutate(vacc_rate_d = vacc_rate - meancov,
         vacc_cat_d = inla.group(vacc_rate_d, n = 10, method = "quantile")) |> 
  left_join(dpt, by = c("plz", "nyeargroup"))

# add geometries
data_full_sf <- shp |> 
  mutate(idarea = row_number()) |> 
  right_join(data_full, by = "plz") |> 
  mutate(plz = factor(plz))
```

`data_full` has columns:
- `plz_2` (character): id for each 2-digit zipcode
- `yeargroup` (character): label for each period (e.g. "2008-2012")
- `nyeargroup` (numeric): index for each period
- `ipd` (numeric): total number of IPD cases for each area and period
- `ipd_19F` (numeric): number of 19F IPD for each area and period
- `ipd_19A` (numeric): number of 19A IPD cases for each area and period
- ... for all other included serotypes
- `ratio_ipd_19F` (numeric): ipd_19F/ipd
- `ratio_ipd_19A` (numeric): ipd_19A/ipd
- ... for all other included serotypes
- `logit_19A` (numeric): log((ratio_ipd_19A+0.001)/(1-ratio_ipd_19A+0.001))
- `vacc_rate` (numeric): PCV vaccination coverage for each area and period (0-1)
- `vacc_cat` (numeric): categorical version of vacc_rate to use as nonlinear effect in INLA (created using inla.group(vacc_rate, n = 10, method = "quantile"))
- `vacc_rate_d` (numeric): delta from mean vaccination coverage for that period
- `vacc_cat_d` (numeric): categorical version of vacc_rate_d
- `dpt_vacc_rate` (numeric): 1st dose DPT vaccination coverage for each area and period (0-1)
- `dpt_vacc_cat` (numeric): categorical version of dpt_vacc_rate

## Visualize data

### Cases

```{r}
case_db <- ipd_raw |> 
  filter(!is.na(DateOfIsolation) & !is.na(agey) & !is.na(Serotype)) |> 
  mutate(ageg = cut(agey, breaks = c(0, 5, 10, 20, 50, 150), include.lowest = TRUE),
         year = year(as.Date(DateOfIsolation, format = "%m/%d/%Y"))) |> 
  group_by(year, ageg, Serotype) |> 
  summarise(ipd_st = n()) |> 
  group_by(year, ageg) |> 
  mutate(ipd_total = sum(ipd_st)) |> 
  ungroup()

case_db |> 
  ggplot(aes(x = year, y = ipd_total, color = ageg)) +
  geom_line() +
  labs(x = "", y = "Total IPD cases", color = "Age group") +
  theme_bw()

case_db |> 
  filter(Serotype == "19F") |> 
  ggplot(aes(x = year, y = ipd_st, color = ageg)) +
  geom_line() +
  labs(x = "", y = "19F IPD cases", color = "Age group") +
  theme_bw()
```

### Scatterplots

```{r, fig.width = 11, fig.height = 4}
# 19F vs % vaccinated
data_full_sf |> 
  ggplot(aes(x = vacc_rate, y = ratio_ipd_19F)) +
  geom_point(aes(size = ipd), alpha = 0.5, shape = 16) +
  #geom_smooth() +
  labs(size = "Total IPD cases") +
  facet_grid(~yeargroup, scales = "free") +
  theme_bw() +
  labs(x = "% vaccinated (PCV)", y = "Proportion 19F cases") +
  theme(text = element_text(size = 16),
        legend.position = "none")

# 19F vs % DPT vaccinated
data_full_sf |> 
  ggplot(aes(x = dpt_vacc_rate, y = ratio_ipd_19F)) +
  geom_point(aes(size = ipd), alpha = 0.5, shape = 16) +
  #geom_smooth() +
  labs(size = "Total IPD cases") +
  facet_grid(~yeargroup, scales = "free") +
  theme_bw() +
  labs(x = "% vaccinated (DPT 1st dose)", y = "Proportion 19F cases") +
  theme(text = element_text(size = 16),
        legend.position = "none")

# 19F vs 19A
data_full_sf |> 
  ggplot(aes(x = ratio_ipd_19A, y = ratio_ipd_19F)) +
  geom_point(aes(size = ipd), alpha = 0.5, shape = 16) +
  #geom_smooth() +
  labs(size = "Total IPD cases") +
  facet_grid(~yeargroup, scales = "free") +
  theme_bw() +
  labs(x = "Proportion 19A cases", y = "Proportion 19F cases") +
  theme(text = element_text(size = 16),
        legend.position = "none")
```

### Univariate maps

```{r, fig.width = 11}
# All IPD cases
data_full_sf |> 
  ggplot() +
  geom_sf(data = shp, fill = "lightgray", color = NA) +
  geom_sf(aes(fill = ipd), color = NA) +
  scale_fill_viridis_c(option = "C") +
  facet_wrap(~yeargroup) +
  labs(fill = "Reported IPD cases") +
  theme_void() +
  theme(text = element_text(size = 16))

# Proportion of 19F cases
data_full_sf |> 
  ggplot() +
  geom_sf(data = shp, fill = "lightgray", color = NA) +
  geom_sf(aes(fill = ratio_ipd_19F), color = NA) +
  scale_fill_viridis_c(option = "C") +
  facet_wrap(~yeargroup) +
  labs(fill = "Proportion\n19F cases\n") +
  theme_void() +
  theme(text = element_text(size = 16))

# Proportion of 19A cases
data_full_sf |> 
  ggplot() +
  geom_sf(data = shp, fill = "lightgray", color = NA) +
  geom_sf(aes(fill = ratio_ipd_19A), color = NA) +
  scale_fill_viridis_c(option = "C") +
  facet_wrap(~yeargroup) +
  labs(fill = "Proportion\n19A cases") +
  theme_void() +
  theme(text = element_text(size = 16))

# PCV vaccine coverage
data_full_sf |> 
  ggplot() +
  geom_sf(data = shp, fill = "lightgray", color = NA) +
  geom_sf(aes(fill = vacc_rate), color = NA) +
  scale_fill_viridis_c(option = "C") +
  facet_wrap(~yeargroup) +
  labs(fill = "PCV coverage\nat 24 months") +
  theme_void() +
  theme(text = element_text(size = 16))

# DPT vaccine coverage
data_full_sf |> 
  ggplot() +
  geom_sf(data = shp, fill = "lightgray", color = NA) +
  geom_sf(aes(fill = dpt_vacc_rate), color = NA) +
  scale_fill_viridis_c(option = "C") +
  facet_wrap(~yeargroup) +
  labs(fill = "DPT 1st dose\ncoverage at\n15 months") +
  theme_void() +
  theme(text = element_text(size = 16))
```

### Bivariate maps

(using Jenks-Fisher algorithm for breaks for the entire period)

```{r, fig.width = 11}
# 19F vs % PCV vaccinated
biv_dat1 <- bi_class(data_full_sf, x = vacc_rate, y = ratio_ipd_19F, style = "fisher", dim = 3)

biv_map1 <- biv_dat1 |> 
  ggplot() +
  geom_sf(aes(fill = bi_class), color = NA, show.legend = FALSE) +
  geom_sf(data = shp, fill = NA, color = "black") +
  bi_scale_fill(pal = "GrPink", dim = 3) +
  facet_wrap(~yeargroup) +
  theme_void() +
  theme(strip.text = element_text(size = 14))

biv_leg1 <- bi_legend(pal = "GrPink",
                     dim = 3,
                     xlab = "Higher PCV vacc ",
                     ylab = "Higher 19F ",
                     size = 12)

plot_grid(biv_map1, biv_leg1, ncol = 2, rel_widths = c(1, 0.25))

# 19F vs % DPT vaccinated
biv_dat2 <- bi_class(data_full_sf, x = dpt_vacc_rate, y = ratio_19F, style = "fisher", dim = 3)

biv_map2 <- biv_dat2 |> 
  ggplot() +
  geom_sf(aes(fill = bi_class), color = NA, show.legend = FALSE) +
  geom_sf(data = shp, fill = NA, color = "black") +
  bi_scale_fill(pal = "GrPink", dim = 3) +
  facet_wrap(~yeargroup) +
  theme_void() +
  theme(strip.text = element_text(size = 14))

biv_leg2 <- bi_legend(pal = "GrPink",
                     dim = 3,
                     xlab = "Higher DPT vacc ",
                     ylab = "Higher 19F ",
                     size = 12)

plot_grid(biv_map2, biv_leg2, ncol = 2, rel_widths = c(1, 0.25))

# 19F vs 19A
biv_dat3 <- bi_class(data_full_sf, x = ratio_ipd_19A, y = ratio_ipd_19F, style = "fisher", dim = 3)

biv_map3 <- biv_dat3 |> 
  ggplot() +
  geom_sf(aes(fill = bi_class), color = NA, show.legend = FALSE) +
  geom_sf(data = shp, fill = NA, color = "black") +
  bi_scale_fill(pal = "GrPink", dim = 3) +
  facet_wrap(~yeargroup) +
  theme_void() +
  theme(strip.text = element_text(size = 14))

biv_leg3 <- bi_legend(pal = "GrPink",
                      dim = 3,
                      xlab = "Higher 19A ",
                      ylab = "Higher 19F ",
                      size = 12)

plot_grid(biv_map3, biv_leg3, ncol = 2, rel_widths = c(1, 0.25))
```

```{r, fig.width = 11}
custom_pal <- c(
  "1-1" = "#9e3547", # low x, low y
  "2-1" = "#70577B",
  "3-1" = "#4279B0", # high x, low y
  "1-2" = "#70577B",
  "2-2" = "#9E3547", # medium x, medium y
  "3-2" = "#70577B",
  "1-3" = "#4279b0", # low x, high y
  "2-3" = "#70577B",
  "3-3" = "#9e3547" # high x, high y
)

biv_map7 <- biv_dat1 |> 
  ggplot() +
  geom_sf(aes(fill = bi_class), color = NA, show.legend = FALSE) +
  geom_sf(data = shp, fill = NA, color = "black") +
  bi_scale_fill(pal = custom_pal, dim = 3) +
  facet_wrap(~yeargroup) +
  theme_void() +
  theme(strip.text = element_text(size = 14))

biv_leg7 <- bi_legend(pal = custom_pal,
                     dim = 3,
                     xlab = "Higher PCV vacc ",
                     ylab = "Higher 19F ",
                     size = 12)

plot_grid(biv_map7, biv_leg7, ncol = 2, rel_widths = c(1, 0.25))

# 19F vs % DPT vaccinated
biv_map8 <- biv_dat2 |> 
  ggplot() +
  geom_sf(aes(fill = bi_class), color = NA, show.legend = FALSE) +
  geom_sf(data = shp, fill = NA, color = "black") +
  bi_scale_fill(pal = custom_pal, dim = 3) +
  facet_wrap(~yeargroup) +
  theme_void() +
  theme(strip.text = element_text(size = 14))

biv_leg8 <- bi_legend(pal = custom_pal,
                     dim = 3,
                     xlab = "Higher DPT vacc ",
                     ylab = "Higher 19F ",
                     size = 12)

plot_grid(biv_map8, biv_leg8, ncol = 2, rel_widths = c(1, 0.25))

# 19F vs 19A
biv_map9 <- biv_dat3 |> 
  ggplot() +
  geom_sf(aes(fill = bi_class), color = NA, show.legend = FALSE) +
  geom_sf(data = shp, fill = NA, color = "black") +
  bi_scale_fill(pal = custom_pal, dim = 3) +
  facet_wrap(~yeargroup) +
  theme_void() +
  theme(strip.text = element_text(size = 14))

biv_leg9 <- bi_legend(pal = custom_pal,
                      dim = 3,
                      xlab = "Higher 19A ",
                      ylab = "Higher 19F ",
                      size = 12)

plot_grid(biv_map9, biv_leg9, ncol = 2, rel_widths = c(1, 0.25))
```

(Fisher-Jenks for each period separately)

```{r, fig.width = 11}
# 19F vs % PCV vaccinated
biv_dat4 <- data_full_sf |> 
  mutate(inv_19F = 1-ratio_ipd_19F) |> 
  group_by(yeargroup) |> 
  mutate(quant_x = cut(vacc_rate, breaks = cartography::getBreaks(vacc_rate, method = "fisher", nclass = 3), include.lowest = TRUE, labels = c(1, 2, 3)),
         quant_y = cut(ratio_ipd_19F, breaks = cartography::getBreaks(ratio_ipd_19F, method = "fisher", nclass = 3), include.lowest = TRUE, labels = c(3, 2, 1)),
         bi_class = paste(quant_x, quant_y, sep = "-"))

biv_map4 <- biv_dat4 |> 
  ggplot() +
  geom_sf(aes(fill = bi_class), color = NA, show.legend = FALSE) +
  geom_sf(data = shp, fill = NA, color = "black") +
  bi_scale_fill(pal = "GrPink", dim = 3) +
  facet_wrap(~yeargroup) +
  theme_void() +
  theme(strip.text = element_text(size = 14))

biv_leg4 <- bi_legend(pal = "GrPink",
                     dim = 3,
                     xlab = "Higher PCV vacc ",
                     ylab = "Lower 19F ",
                     size = 12)

plot_grid(biv_map4, biv_leg4, ncol = 2, rel_widths = c(1, 0.25))

# 19F vs % DPT vaccinated
biv_dat5 <- data_full_sf |> 
  mutate(inv_19F = 1-ratio_ipd_19F) |> 
  group_by(yeargroup) |> 
  mutate(quant_x = cut(dpt_vacc_rate, breaks = cartography::getBreaks(dpt_vacc_rate, method = "fisher", nclass = 3), include.lowest = TRUE, labels = c(1, 2, 3)),
         quant_y = cut(ratio_ipd_19F, breaks = cartography::getBreaks(ratio_ipd_19F, method = "fisher", nclass = 3), include.lowest = TRUE, labels = c(3, 2, 1)),
         bi_class = paste(quant_x, quant_y, sep = "-"))

biv_map5 <- biv_dat5 |> 
  ggplot() +
  geom_sf(aes(fill = bi_class), color = NA, show.legend = FALSE) +
  geom_sf(data = shp, fill = NA, color = "black") +
  bi_scale_fill(pal = "GrPink", dim = 3) +
  facet_wrap(~yeargroup) +
  theme_void() +
  theme(strip.text = element_text(size = 14))

biv_leg5 <- bi_legend(pal = "GrPink",
                     dim = 3,
                     xlab = "Higher DPT vacc ",
                     ylab = "Lower 19F ",
                     size = 12)

plot_grid(biv_map5, biv_leg5, ncol = 2, rel_widths = c(1, 0.25))

# 19F vs 19A
biv_dat6 <- data_full_sf |> 
  mutate(inv_19F = 1-ratio_ipd_19F) |> 
  group_by(yeargroup) |> 
  mutate(quant_x = cut(ratio_ipd_19A, breaks = cartography::getBreaks(ratio_ipd_19A, method = "fisher", nclass = 3), include.lowest = TRUE, labels = c(1, 2, 3)),
         quant_y = cut(ratio_ipd_19F, breaks = cartography::getBreaks(ratio_ipd_19F, method = "fisher", nclass = 3), include.lowest = TRUE, labels = c(3, 2, 1)),
         bi_class = paste(quant_x, quant_y, sep = "-"))

biv_map6 <- biv_dat6 |> 
  ggplot() +
  geom_sf(aes(fill = bi_class), color = NA, show.legend = FALSE) +
  geom_sf(data = shp, fill = NA, color = "black") +
  bi_scale_fill(pal = "GrPink", dim = 3) +
  facet_wrap(~yeargroup) +
  theme_void() +
  theme(strip.text = element_text(size = 14))

biv_leg6 <- bi_legend(pal = "GrPink",
                      dim = 3,
                      xlab = "Higher 19A ",
                      ylab = "Lower 19F ",
                      size = 12)

plot_grid(biv_map6, biv_leg6, ncol = 2, rel_widths = c(1, 0.25))
```

```{r, fig.width = 11}
# 19F vs % PCV vaccinated
biv_map10 <- biv_dat4 |> 
  ggplot() +
  geom_sf(aes(fill = bi_class), color = NA, show.legend = FALSE) +
  geom_sf(data = shp, fill = NA, color = "black") +
  bi_scale_fill(pal = custom_pal, dim = 3) +
  facet_wrap(~yeargroup) +
  theme_void() +
  theme(strip.text = element_text(size = 14))

biv_leg10 <- bi_legend(pal = custom_pal,
                     dim = 3,
                     xlab = "Higher PCV vacc ",
                     ylab = "Lower 19F ",
                     size = 12)

plot_grid(biv_map10, biv_leg10, ncol = 2, rel_widths = c(1, 0.25))

# 19F vs % DPT vaccinated
biv_map11 <- biv_dat5 |> 
  ggplot() +
  geom_sf(aes(fill = bi_class), color = NA, show.legend = FALSE) +
  geom_sf(data = shp, fill = NA, color = "black") +
  bi_scale_fill(pal = custom_pal, dim = 3) +
  facet_wrap(~yeargroup) +
  theme_void() +
  theme(strip.text = element_text(size = 14))

biv_leg11 <- bi_legend(pal = custom_pal,
                     dim = 3,
                     xlab = "Higher DPT vacc ",
                     ylab = "Lower 19F ",
                     size = 12)

plot_grid(biv_map11, biv_leg11, ncol = 2, rel_widths = c(1, 0.25))

# 19F vs 19A
biv_map12 <- biv_dat6 |> 
  ggplot() +
  geom_sf(aes(fill = bi_class), color = NA, show.legend = FALSE) +
  geom_sf(data = shp, fill = NA, color = "black") +
  bi_scale_fill(pal = custom_pal, dim = 3) +
  facet_wrap(~yeargroup) +
  theme_void() +
  theme(strip.text = element_text(size = 14))

biv_leg12 <- bi_legend(pal = custom_pal,
                      dim = 3,
                      xlab = "Higher 19A ",
                      ylab = "Lower 19F ",
                      size = 12)

plot_grid(biv_map12, biv_leg12, ncol = 2, rel_widths = c(1, 0.25))
```

## Models

### Create adjacency matrix

```{r}
# nb <- poly2nb(shp, queen = TRUE)
# nb2INLA("map.adj", nb)
adj_mat <- inla.read.graph(filename = "map.adj")
```

### Define functions for running models and plotting

```{r}
# run model
runmod <- function(data, 
                   cases_19F = "ipd_19F", 
                   cases_19A = "ipd_19A", 
                   ratio_ipd_19A = "ratio_ipd_19A",
                   ratio_ipd_23F = "ratio_ipd_23F",
                   ratio_ipd_14 = "ratio_ipd_14",
                   ratio_ipd_18C = "ratio_ipd_18C",
                   ratio_ipd_6B = "ratio_ipd_6B",
                   ratio_ipd_9V = "ratio_ipd_9V",
                   ratio_ipd_4 = "ratio_ipd_4",
                   ratio_ipd_pcv7 = "ratio_ipd_pcv7",
                   logit_ratio_ipd_19A = "logit_ratio_ipd_19A",
                   logit_ratio_ipd_23F = "logit_ratio_ipd_23F",
                   logit_ratio_ipd_14 = "logit_ratio_ipd_14",
                   logit_ratio_ipd_18C = "logit_ratio_ipd_18C",
                   logit_ratio_ipd_6B = "logit_ratio_ipd_6B",
                   logit_ratio_ipd_9V = "logit_ratio_ipd_9V",
                   logit_ratio_ipd_4 = "logit_ratio_ipd_4",
                   logit_ratio_ipd_pcv7 = "logit_ratio_ipd_pcv7",
                   N = "ipd", 
                   idarea = "idarea", 
                   vacc_rate = "vacc_rate", 
                   vacc_cat = "vacc_cat",
                   vacc_rate_d = "vacc_rate_d",
                   vacc_cat_d = "vacc_cat_d",
                   dpt_vacc_rate = "dpt_vacc_rate",
                   dpt_vacc_cat = "dpt_vacc_cat",
                   t = "nyeargroup", 
                   formula) {
  
  data_inla <- list(
    cases_19F = data[[cases_19F]],
    cases_19A = data[[cases_19A]],
    ratio_ipd_19A = data[[ratio_ipd_19A]],
    ratio_ipd_23F = data[[ratio_ipd_23F]],
    ratio_ipd_14 = data[[ratio_ipd_14]],
    ratio_ipd_18C = data[[ratio_ipd_18C]],
    ratio_ipd_6B = data[[ratio_ipd_6B]],
    ratio_ipd_9V = data[[ratio_ipd_9V]],
    ratio_ipd_4 = data[[ratio_ipd_4]],
    ratio_ipd_pcv7 = data[[ratio_ipd_pcv7]],
    logit_ratio_ipd_19A = data[[logit_ratio_ipd_19A]],
    logit_ratio_ipd_23F = data[[logit_ratio_ipd_23F]],
    logit_ratio_ipd_14 = data[[logit_ratio_ipd_14]],
    logit_ratio_ipd_18C = data[[logit_ratio_ipd_18C]],
    logit_ratio_ipd_6B = data[[logit_ratio_ipd_6B]],
    logit_ratio_ipd_9V = data[[logit_ratio_ipd_9V]],
    logit_ratio_ipd_4 = data[[logit_ratio_ipd_4]],
    logit_ratio_ipd_pcv7 = data[[logit_ratio_ipd_pcv7]],
    total_cases = data[[N]],
    N = data[[N]],
    idarea = data[[idarea]],
    vacc_rate = data[[vacc_rate]],
    vacc_cat = data[[vacc_cat]],
    vacc_rate_d = data[[vacc_rate_d]],
    vacc_cat_d = data[[vacc_cat_d]],
    dpt_vacc_rate = data[[dpt_vacc_rate]],
    dpt_vacc_cat = data[[dpt_vacc_cat]],
    t = data[[t]]
  )
  
  
  model_result <- inla(
    as.formula(formula), 
    family = "binomial",
    data = data_inla, 
    Ntrials = data_inla$N, 
    control.predictor = list(compute = TRUE), 
    control.compute = list(dic = TRUE, waic = TRUE),
    verbose = TRUE
  )
  
  return(model_result)
}

# plot bym
plot_bym <- function(result, shapefile) {
  n <- nrow(result$summary.random$idarea)
  split_point <- n / 2
  
  bym_result <- shapefile |> 
    cbind(result$summary.random$idarea[1:split_point, ]) |> 
    mutate(sig = case_when(X0.025quant > 0 & X0.975quant > 0 ~ 1,
                           X0.025quant < 0 & X0.975quant < 0 ~ 1,
                           .default = 0))
  
  bym_result |> 
    ggplot() +
    geom_sf(aes(fill = exp(mean)), color = NA) +
    scale_fill_viridis_c() +
    labs(fill = "Effect") +
    theme_void()
}

# plot ar1
plot_ar1 <- function(result, time_mapping) {
  ar1_result <- result$summary.random$t |> 
    left_join(time_mapping, by = c("ID" = "nyeargroup")) |> 
    rename(lb = `0.025quant`, ub = `0.975quant`)
  
  ar1_result |> 
    ggplot(aes(x = yeargroup, color = yeargroup)) +
    geom_errorbar(aes(ymin = exp(lb), ymax = exp(ub)), show.legend = FALSE) +
    geom_point(aes(y = exp(mean)), show.legend = FALSE) +
    theme_bw() +
    labs(x = "", y = "Effect")
}

# plot linear predictors
plot_lp <- function(result, var_name, data, label = NULL) {
  fe <- result$summary.fixed[var_name, ]
  
  predictor_values <- seq(min(data[[var_name]], na.rm = TRUE), max(data[[var_name]], na.rm = TRUE), length.out = 100)
  
  effect_df <- data.frame(predictor = predictor_values,
                          effect = exp(fe$mean * predictor_values),
                          lower = exp(fe$`0.025quant` * predictor_values),
                          upper = exp(fe$`0.975quant` * predictor_values))

  effect_df |> 
    ggplot(aes(x = predictor, y = effect)) +
    #geom_hline(aes(yintercept = 1), lty = 2, color = "red") +
    geom_line(linewidth = 1) +
    geom_ribbon(aes(ymin = lower, ymax = upper), alpha = 0.2) +
    labs(x = label %||% predictor_name, y = "Effect") +
    theme_bw()
}

# plot nonlinear predictors
plot_nlp <- function(result, var_name, label = NULL) {
  nonlinear_result <- result$summary.random[[var_name]]
  
  ggplot(nonlinear_result, aes(x = ID)) +
    #geom_hline(aes(yintercept = 1), lty = 2, color = "red") +
    geom_errorbar(aes(ymin = exp(`0.025quant`), ymax = exp(`0.975quant`))) +
    geom_point(aes(y = exp(mean)), size = 2) +
    labs(
      x = label %||% var_name,
      y = "Effect"
    ) +
    theme_bw()
}
```

### Baseline model

```{r}
base_f <- "cases_19F ~ 1 + f(idarea, model = 'bym2', graph = adj_mat) + f(t, model = 'ar1')"

base_mod <- runmod(data = data_full_sf, formula = base_f)

summary(base_mod)
plot_bym(result = base_mod, shapefile = shp)
plot_ar1(result = base_mod, time_mapping = data_full |> dplyr::select(nyeargroup, yeargroup) |> distinct())
```

### Vaccine coverage (linear)

```{r}
f2 <- "cases_19F ~ 1 + f(idarea, model = 'bym2', graph = adj_mat) + f(t, model = 'ar1') + vacc_rate"

mod2 <- runmod(data = data_full_sf, formula = f2)

summary(mod2)
plot_bym(result = mod2, shapefile = shp)
plot_ar1(result = mod2, time_mapping = data_full |> dplyr::select(nyeargroup, yeargroup) |> distinct())
plot_lp(data = data_full, result = mod2, var_name = "vacc_rate", label = "PCV coverage")
```

### Vaccine coverage (nonlinear, rw1)

```{r}
# Extra constraint to fix lowest category of vaccines as reference
ec_f3 <- list(A = matrix(0, nrow = 1, ncol = length(unique((data_full$vacc_cat)))), e = 0)
ec_f3$A[1, 1] <- 1

f3 <- "cases_19F ~ 1 + f(idarea, model = 'bym2', graph = adj_mat) + f(t, model = 'ar1') + f(vacc_cat, model = 'rw1', constr = FALSE, extraconstr = ec_f3)"

mod3 <- runmod(data = data_full_sf, formula = f3)

summary(mod3)
plot_bym(result = mod3, shapefile = shp)
plot_ar1(result = mod3, time_mapping = data_full |> dplyr::select(nyeargroup, yeargroup) |> distinct())
plot_nlp(result = mod3, var_name = "vacc_cat", label = "PCV coverage")
```

### Delta vaccine coverage (linear)

```{r}
f4 <- "cases_19F ~ 1 + f(idarea, model = 'bym2', graph = adj_mat) + f(t, model = 'ar1') + vacc_rate_d"

mod4 <- runmod(data = data_full_sf, formula = f4)

summary(mod4)
plot_bym(result = mod4, shapefile = shp)
plot_ar1(result = mod4, time_mapping = data_full |> dplyr::select(nyeargroup, yeargroup) |> distinct())
plot_lp(data = data_full, result = mod4, var_name = "vacc_rate_d", label = "Delta PCV coverage")
```

### DPT vaccine coverage (linear)

```{r}
f5 <- "cases_19F ~ 1 + f(idarea, model = 'bym2', graph = adj_mat) + f(t, model = 'ar1') + dpt_vacc_rate"

mod5 <- runmod(data = data_full_sf, formula = f5)

summary(mod5)
plot_bym(result = mod5, shapefile = shp)
plot_ar1(result = mod5, time_mapping = data_full |> dplyr::select(nyeargroup, yeargroup) |> distinct())
plot_lp(data = data_full, result = mod5, var_name = "dpt_vacc_rate", label = "DPT coverage")
```

### 19A proportion (linear)

```{r}
f6 <- "cases_19F ~ 1 + f(idarea, model = 'bym2', graph = adj_mat) + f(t, model = 'ar1') + ratio_ipd_19A"

mod6 <- runmod(data = data_full_sf, formula = f6)

summary(mod6)
plot_bym(result = mod6, shapefile = shp)
plot_ar1(result = mod6, time_mapping = data_full |> dplyr::select(nyeargroup, yeargroup) |> distinct())
plot_lp(data = data_full, result = mod6, var_name = "ratio_ipd_19A", label = "Proportion 19A")
```

```{r}
f7 <- "cases_19F ~ 1 + f(idarea, model = 'bym2', graph = adj_mat) + f(t, model = 'ar1') + logit_ratio_ipd_19A"

mod7 <- runmod(data = data_full_sf, formula = f7)

summary(mod7)
plot_bym(result = mod7, shapefile = shp)
plot_ar1(result = mod7, time_mapping = data_full |> dplyr::select(nyeargroup, yeargroup) |> distinct())
plot_lp(data = data_full, result = mod7, var_name = "logit_ratio_ipd_19A", label = "logit(Proportion 19A)")

data.frame(predictor = seq(min(data_full$logit_ratio_ipd_19A, na.rm = TRUE), max(data_full$logit_ratio_ipd_19A, na.rm = TRUE), length.out = 100)) |> 
  mutate(effect = 1 / (1 + exp(-mod7$summary.fixed["logit_ratio_ipd_19A", ]$mean * predictor)),
         lower = 1/ (1 + exp(-mod7$summary.fixed["logit_ratio_ipd_19A", ]$`0.025quant` * predictor)),
         upper = 1/ (1 + exp(-mod7$summary.fixed["logit_ratio_ipd_19A", ]$`0.975quant` * predictor))) |> 
    ggplot(aes(x = exp(predictor), y = effect)) +
    geom_line(size = 1) +
    geom_ribbon(aes(ymin = lower, ymax = upper), alpha = 0.2) +
    geom_hline(aes(yintercept = 1), lty = 2, color = "red") +
    labs(x = "Proportion 19A", y = "Effect") +
    theme_bw()
```

### Other PCV7 serotypes

```{r}
f8 <- "cases_19F ~ 1 + f(idarea, model = 'bym2', graph = adj_mat) + f(t, model = 'ar1') + logit_ratio_ipd_pcv7"

mod8 <- runmod(data = data_full_sf, formula = f8)

summary(mod8)
plot_bym(result = mod8, shapefile = shp)
plot_ar1(result = mod8, time_mapping = data_full |> dplyr::select(nyeargroup, yeargroup) |> distinct())
plot_lp(data = data_full, result = mod8, var_name = "logit_ratio_ipd_pcv7", label = "Proportion non-19F PCV7 sts")
```

## Compare WAIC/DICs

```{r}
mods <- list(base_mod = base_mod, mod2 = mod2, mod3 = mod3, mod4 = mod4, mod5 = mod5, mod6 = mod6, mod7 = mod7, mod8 = mod8)

mod_comp <- data.frame(model = names(mods),
                       formula = c(base_f, f2, f3, f4, f5, f6, f7, f8),
                       waic = sapply(mods, function(res) round(res$waic$waic, 2)),
                       dic = sapply(mods, function(res) round(res$dic$dic, 2))) |> 
  mutate(waic_delta = waic - waic[1],
         dic_delta = dic - dic[1])
```

## Only with >= 65 years old

```{r}
data_full65 <- ipd_raw |> 
  filter(agey >= 65) |> 
  mutate(date = as.Date(DateOfIsolation, format = "%m/%d/%Y"),
         plz_pre = str_pad(PLZpatient, side = "left", pad = "0", width = 3),
         plz = substr(plz_pre, 1, 2),
         year = year(date)) |> 
  filter(!is.na(PLZpatient) & date > as.Date("2009-01-01") & !is.na(date) & !(plz %in% c("05", "11"))) |> 
  arrange(year) |> 
  group_by(year) |> 
  mutate(nyear = cur_group_id()) |> 
  ungroup() |> 
  mutate(yeargroup = factor(cut(nyear, 3, labels = FALSE), levels = c(1:3), labels = c("2008-2012", "2013-2017", "2018-2022")),
         nyeargroup = cut(nyear, 3, labels = FALSE)) |> 
  group_by(plz, Serotype, yeargroup, nyeargroup) |> 
  summarise(ipd_st = n()) |> 
  group_by(plz, yeargroup, nyeargroup) |> 
  mutate(ipd = sum(ipd_st),
         ipd_pcv7 = sum(ipd_st[Serotype %in% pcv7_no19f], na.rm = TRUE)) |> 
  ungroup() |> 
  filter(Serotype %in% pcv7_st | Serotype == "19A") |> 
  pivot_wider(names_from = Serotype, names_prefix = "ipd_", values_from = ipd_st) |> 
  mutate(across(starts_with("ipd_"), ~ if_else(is.na(.x), 0, .x)), 
         across(starts_with("ipd_"), ~ .x / ipd, .names = "ratio_{.col}"),
         across(starts_with("ratio_ipd_"), ~ log((.x + 0.001) / (1 - .x + 0.001)), .names = "logit_{.col}")) |> 
  left_join(vax_cov_agg, by = c("plz", "nyeargroup")) |> 
  group_by(yeargroup) |> 
  mutate(meancov = mean(vacc_rate, na.rm = TRUE)) |> 
  ungroup() |> 
  mutate(vacc_rate_d = vacc_rate - meancov,
         vacc_cat_d = inla.group(vacc_rate_d, n = 10, method = "quantile")) |> 
  left_join(dpt, by = c("plz", "nyeargroup"))

# add geometries
data_full65_sf <- shp |> 
  mutate(idarea = row_number()) |> 
  right_join(data_full, by = "plz") |> 
  mutate(plz = factor(plz))
```

### Baseline

```{r}
base_f <- "cases_19F ~ 1 + f(idarea, model = 'bym2', graph = adj_mat) + f(t, model = 'ar1')"

base_mod65 <- runmod(data = data_full65_sf, formula = base_f)

summary(base_mod65)
plot_bym(result = base_mod65, shapefile = shp)
plot_ar1(result = base_mod65, time_mapping = data_full65 |> dplyr::select(nyeargroup, yeargroup) |> distinct())
```

### Vaccine coverage (linear)

```{r}
f2 <- "cases_19F ~ 1 + f(idarea, model = 'bym2', graph = adj_mat) + f(t, model = 'ar1') + vacc_rate"

mod2_65 <- runmod(data = data_full65_sf, formula = f2)

summary(mod2_65)
plot_bym(result = mod2_65, shapefile = shp)
plot_ar1(result = mod2_65, time_mapping = data_full65 |> dplyr::select(nyeargroup, yeargroup) |> distinct())
plot_lp(data = data_full65, result = mod2_65, var_name = "vacc_rate", label = "PCV coverage")
```

### Delta vaccine coverage (linear)

```{r}
f4 <- "cases_19F ~ 1 + f(idarea, model = 'bym2', graph = adj_mat) + f(t, model = 'ar1') + vacc_rate_d"

mod4_65 <- runmod(data = data_full65_sf, formula = f4)

summary(mod4_65)
plot_bym(result = mod4_65, shapefile = shp)
plot_ar1(result = mod4_65, time_mapping = data_full65 |> dplyr::select(nyeargroup, yeargroup) |> distinct())
plot_lp(data = data_full65, result = mod4_65, var_name = "vacc_rate_d", label = "Delta PCV coverage")
```

### DPT vaccine coverage (linear)

```{r}
f5 <- "cases_19F ~ 1 + f(idarea, model = 'bym2', graph = adj_mat) + f(t, model = 'ar1') + dpt_vacc_rate"

mod5_65 <- runmod(data = data_full65_sf, formula = f5)

summary(mod5_65)
plot_bym(result = mod5_65, shapefile = shp)
plot_ar1(result = mod5_65, time_mapping = data_full65 |> dplyr::select(nyeargroup, yeargroup) |> distinct())
plot_lp(data = data_full65, result = mod5_65, var_name = "dpt_vacc_rate", label = "DPT coverage")
```

### 19A proportion (linear)

```{r}
f6 <- "cases_19F ~ 1 + f(idarea, model = 'bym2', graph = adj_mat) + f(t, model = 'ar1') + ratio_ipd_19A"

mod6_65 <- runmod(data = data_full65_sf, formula = f6)

summary(mod6_65)
plot_bym(result = mod6_65, shapefile = shp)
plot_ar1(result = mod6_65, time_mapping = data_full65 |> dplyr::select(nyeargroup, yeargroup) |> distinct())
plot_lp(data = data_full65, result = mod6_65, var_name = "ratio_ipd_19A", label = "Proportion 19A")
```

```{r}
f7 <- "cases_19F ~ 1 + f(idarea, model = 'bym2', graph = adj_mat) + f(t, model = 'ar1') + logit_ratio_ipd_19A"

mod7_65 <- runmod(data = data_full65_sf, formula = f7)

summary(mod7_65)
plot_bym(result = mod7_65, shapefile = shp)
plot_ar1(result = mod7_65, time_mapping = data_full65 |> dplyr::select(nyeargroup, yeargroup) |> distinct())
plot_lp(data = data_full65, result = mod7_65, var_name = "logit_ratio_ipd_19A", label = "logit(Proportion 19A)")
```

### Other PCV7 serotypes

```{r}
f8 <- "cases_19F ~ 1 + f(idarea, model = 'bym2', graph = adj_mat) + f(t, model = 'ar1') + logit_ratio_ipd_pcv7"

mod8_65 <- runmod(data = data_full65_sf, formula = f8)

summary(mod8_65)
plot_bym(result = mod8_65, shapefile = shp)
plot_ar1(result = mod8_65, time_mapping = data_full65 |> dplyr::select(nyeargroup, yeargroup) |> distinct())
plot_lp(data = data_full65, result = mod8_65, var_name = "logit_ratio_ipd_pcv7", label = "Proportion non-19F PCV7 sts")
```
